+++ containers/src/Data/Tree.hs
@@ -33,6 +33,7 @@ module Data.Tree(
     -- * Trees and Forests
       Tree(..)
     , Forest
+    , PostOrder(..)
 
     -- * Construction
     , unfoldTree
@@ -46,6 +47,10 @@ module Data.Tree(
     , foldTree
     , flatten
     , levels
+    , leaves
+    , edges
+    , pathsToRoot
+    , pathsFromRoot
 
     -- * Ascii Drawings
     , drawTree
@@ -55,7 +60,10 @@ module Data.Tree(
 
 import Utils.Containers.Internal.Prelude as Prelude
 import Prelude ()
-import Data.Foldable (fold, toList)
+import Data.Bits ((.&.))
+import Data.Foldable (toList)
+import qualified Data.Foldable as Foldable
+import Data.List.NonEmpty (NonEmpty(..))
 import Data.Traversable (foldMapDefault)
 import Control.Monad (liftM)
 import Control.Monad.Fix (MonadFix (..), fix)
@@ -66,6 +74,7 @@ import Control.DeepSeq (NFData(rnf),NFData1(liftRnf))
 #ifdef __GLASGOW_HASKELL__
 import Data.Data (Data)
 import GHC.Generics (Generic, Generic1)
+import qualified GHC.Exts
 import Language.Haskell.TH.Syntax (Lift)
 -- See Note [ Template Haskell Dependencies ]
 import Language.Haskell.TH ()
@@ -74,7 +83,7 @@ import Language.Haskell.TH ()
 import Control.Monad.Zip (MonadZip (..))
 
 #ifdef __GLASGOW_HASKELL__
-import Data.Coerce
+import Data.Coerce (coerce)
 #endif
 import Data.Functor.Classes
 
@@ -84,7 +93,6 @@ import Data.Semigroup (Semigroup (..))
 
 #if MIN_VERSION_base(4,18,0)
 import qualified Data.Foldable1 as Foldable1
-import Data.List.NonEmpty (NonEmpty(..))
 #endif
 
 -- | Non-empty, possibly infinite, multi-way trees; also known as /rose trees/.
@@ -184,12 +192,13 @@ mfixTree f
   = Node a (zipWith (\i _ -> mfixTree ((!! i) . subForest . f))
                     [0..] children)
 
+-- | Traverses in pre-order.
 instance Traversable Tree where
   traverse f = go
     where go (Node x ts) = liftA2 Node (f x) (traverse go ts)
   {-# INLINE traverse #-}
 
--- | Folds in preorder
+-- | Folds in pre-order.
 
 -- See Note [Implemented Foldable Tree functions]
 instance Foldable Tree where
@@ -234,7 +243,7 @@ instance Foldable Tree where
     {-# INLINABLE product #-}
 
 #if MIN_VERSION_base(4,18,0)
--- | Folds in preorder
+-- | Folds in pre-order.
 --
 -- @since 0.6.7
 
@@ -251,17 +260,17 @@ instance Foldable1.Foldable1 Tree where
       go (Node x []) = f x
       go (Node x (t : ts)) =
         f x <> Foldable1.foldrMap1 go (\t' z -> go t' <> z) (t :| ts)
-  {-# INLINABLE foldMap1 #-}
+  {-# INLINE foldMap1 #-}
 
   foldMap1' f = foldlMap1' f (\z x -> z <> f x)
-  {-# INLINABLE foldMap1' #-}
+  {-# INLINE foldMap1' #-}
 
   toNonEmpty (Node x ts) = x :| concatMap toList ts
 
-  maximum = maximum
+  maximum = Foldable.maximum
   {-# INLINABLE maximum #-}
 
-  minimum = minimum
+  minimum = Foldable.minimum
   {-# INLINABLE minimum #-}
 
   foldrMap1 = foldrMap1
@@ -364,6 +373,8 @@ draw (Node x ts0) = lines x ++ drawSubTrees ts0
     shift first other = zipWith (++) (first : repeat other)
 
 -- | Returns the elements of a tree in pre-order.
+--
+-- @flatten == Data.Foldable.'toList'@
 --
 -- @
 --
@@ -532,6 +543,277 @@ unfoldForestQ f aQ = case viewl aQ of
         q' :> a -> splitOnto (a:as) bs q'
         EmptyR -> error "unfoldForestQ"
 
+-- | \(O(n)\). The leaves of the tree in left-to-right order.
+--
+-- A leaf is a node with no children.
+--
+-- ==== __Examples__
+--
+-- >>> :{
+-- leaves $
+--   Node 1
+--     [ Node 2
+--         [ Node 4 []
+--         , Node 5 []
+--         ]
+--     , Node 3
+--         [ Node 6 []
+--         ]
+--     ]
+-- :}
+-- [4,5,6]
+-- >>> leaves (Node "root" [])
+-- ["root"]
+--
+-- @since FIXME
+leaves :: Tree a -> [a]
+#ifdef __GLASGOW_HASKELL__
+leaves t = GHC.Exts.build $ \cons nil ->
+  let go (Node x []) z = cons x z
+      go (Node _ ts) z = foldr go z ts
+  in go t nil
+{-# INLINE leaves #-} -- Inline for list fusion
+#else
+leaves t =
+  let go (Node x []) z = x:z
+      go (Node _ ts) z = foldr go z ts
+  in go t []
+#endif
+
+-- | \(O(n)\). The edges of the tree as parent-child pairs in pre-order.
+--
+-- A tree with \(n\) nodes has \(n-1\) edges.
+--
+-- ==== __Examples__
+--
+-- >>> :{
+-- edges $
+--   Node 1
+--     [ Node 2
+--         [ Node 4 []
+--         , Node 5 []
+--         ]
+--     , Node 3
+--         [ Node 6 []
+--         ]
+--     ]
+-- :}
+-- [(1,2),(2,4),(2,5),(1,3),(3,6)]
+-- >>> edges (Node "root" [])
+-- []
+--
+-- @since FIXME
+edges :: Tree a -> [(a, a)]
+#ifdef __GLASGOW_HASKELL__
+edges (Node x0 ts0) = GHC.Exts.build $ \cons nil ->
+  let go p = foldr (\(Node x ts) z -> cons (p, x) (go x z ts))
+  in go x0 nil ts0
+{-# INLINE edges #-} -- Inline for list fusion
+#else
+edges (Node x0 ts0) =
+  let go p = foldr (\(Node x ts) z -> (p, x) : go x z ts)
+  in go x0 [] ts0
+#endif
+
+-- | \(O(n)\). Labels on the paths from each node to the root.
+--
+-- ==== __Examples__
+--
+-- >>> :{
+-- pathsToRoot $
+--   Node 1
+--     [ Node 2 []
+--     , Node 3 []
+--     ]
+-- :}
+-- Node {rootLabel = 1 :| [], subForest = [Node {rootLabel = 2 :| [1], subForest = []},Node {rootLabel = 3 :| [1], subForest = []}]}
+-- >>> pathsToRoot (Node "root" [])
+-- Node {rootLabel = "root" :| [], subForest = []}
+--
+-- @since FIXME
+pathsToRoot :: Tree a -> Tree (NonEmpty a)
+pathsToRoot = go []
+  where
+    go ps (Node x ts) = Node (x :| ps) (map (go (x:ps)) ts)
+
+-- | Labels on the paths from the root to each node.
+--
+-- If the path orientation is not important, consider using 'pathsToRoot'
+-- instead because it is more efficient.
+--
+-- ==== __Examples__
+--
+-- >>> :{
+-- pathsFromRoot $
+--   Node 1
+--     [ Node 2 []
+--     , Node 3 []
+--     ]
+-- :}
+-- Node {rootLabel = 1 :| [], subForest = [Node {rootLabel = 1 :| [2], subForest = []},Node {rootLabel = 1 :| [3], subForest = []}]}
+-- >>> pathsFromRoot (Node "root" [])
+-- Node {rootLabel = "root" :| [], subForest = []}
+--
+-- @since FIXME
+
+-- See Note [pathsFromRoot implementation]
+pathsFromRoot :: Tree a -> Tree (NonEmpty a)
+pathsFromRoot (Node x0 ts0) = Node (x0 :| []) (map (go (singletonBQ x0)) ts0)
+  where
+    go !q (Node x ts) = Node (toNonEmptyBQ q') (map (go q') ts)
+      where
+        !q' = snocBQ q x
+
+-- An implementation of Chris Okasaki's banker's queue.
+-- Invariant: length front >= length rear
+data BQ a = BQ
+  a -- head
+  {-# UNPACK #-} !Word -- length front + length rear
+  [a] -- front
+  ![a] -- rear (reversed)
+
+singletonBQ :: a -> BQ a
+singletonBQ x = BQ x 0 [] []
+
+snocBQ :: BQ a -> a -> BQ a
+snocBQ (BQ x0 n f r) x
+  | doReverse = BQ x0 (n+1) (f ++ reverse (x:r)) []
+  | otherwise = BQ x0 (n+1) f (x:r)
+  where
+    doReverse = (n+2) .&. (n+1) == 0
+    -- We reverse whenever the length of r would exceed that of f.
+    -- This happens every time n+2 is a power of 2.
+
+toNonEmptyBQ :: BQ a -> NonEmpty a
+toNonEmptyBQ (BQ x0 _ f r) = case r of
+  [] -> x0 :| f -- optimization, no need to rebuild f
+  _ -> x0 :| (f ++ reverse r)
+
+-- | A newtype over 'Tree' that folds and traverses in post-order.
+--
+-- @since FIXME
+newtype PostOrder a = PostOrder { unPostOrder :: Tree a }
+#ifdef __GLASGOW_HASKELL__
+  deriving (Eq, Ord, Read, Show, Data, Generic, Generic1, Lift)
+#else
+  deriving (Eq, Ord, Read, Show)
+#endif
+
+instance Functor PostOrder where
+#ifdef __GLASGOW_HASKELL__
+  fmap = (coerce :: ((a -> b) -> Tree a -> Tree b)
+                 -> (a -> b) -> PostOrder a -> PostOrder b)
+         fmapTree
+  (<$) = (coerce :: (b -> Tree a -> Tree b)
+                 -> b -> PostOrder a -> PostOrder b)
+         (<$)
+#else
+  fmap f = PostOrder . fmapTree f . unPostOrder
+  (<$) x = PostOrder . (x <$) . unPostOrder
+#endif
+
+-- See Note [Implemented Foldable Tree functions]
+instance Foldable PostOrder where
+    fold = foldMap id
+    {-# INLINABLE fold #-}
+
+    foldMap = foldMapDefault
+    {-# INLINE foldMap #-}
+
+    foldr f z0 = \(PostOrder t) -> go t z0  -- Use a lambda to inline with two arguments
+      where
+        go (Node x ts) z = foldr go (f x z) ts
+    {-# INLINE foldr #-}
+
+    foldl' f z0 = \(PostOrder t) -> go z0 t  -- Use a lambda to inline with two arguments
+      where
+        go !z (Node x ts) =
+          let !z' = foldl' go z ts
+          in f z' x
+    {-# INLINE foldl' #-}
+
+    foldr1 = foldrMap1PostOrder id
+
+    foldl1 = foldlMap1PostOrder id
+
+    null _ = False
+    {-# INLINE null #-}
+
+    elem = any . (==)
+    {-# INLINABLE elem #-}
+
+    maximum = foldlMap1'PostOrder id max
+    {-# INLINABLE maximum #-}
+
+    minimum = foldlMap1'PostOrder id min
+    {-# INLINABLE minimum #-}
+
+    sum = foldlMap1'PostOrder id (+)
+    {-# INLINABLE sum #-}
+
+    product = foldlMap1'PostOrder id (*)
+    {-# INLINABLE product #-}
+
+instance Traversable PostOrder where
+  traverse f = \(PostOrder t) -> PostOrder <$> go t
+    where
+      go (Node x ts) = liftA2 (flip Node) (traverse go ts) (f x)
+  {-# INLINE traverse #-}
+
+#if MIN_VERSION_base(4,18,0)
+-- See Note [Implemented Foldable1 Tree functions]
+instance Foldable1.Foldable1 PostOrder where
+  foldMap1 f = \(PostOrder t) -> go t  -- Use a lambda to inline with one argument
+    where
+      go (Node x []) = f x
+      go (Node x (t:ts)) =
+        Foldable1.foldrMap1 go (\t' z' -> go t' <> z') (t :| ts) <> f x
+  {-# INLINE foldMap1 #-}
+
+  foldMap1' f = foldlMap1'PostOrder f (\z x -> z <> f x)
+  {-# INLINE foldMap1' #-}
+
+  toNonEmpty (PostOrder t0) = go t0 []
+    where
+      go (Node x []) z = x :| z
+      go (Node x (t:ts)) z =
+        go t (foldr (\t' z' -> foldr (:) z' (PostOrder t')) (x:z) ts)
+
+  maximum = Foldable.maximum
+  {-# INLINABLE maximum #-}
+
+  minimum = Foldable.minimum
+  {-# INLINABLE minimum #-}
+
+  foldrMap1 = foldrMap1PostOrder
+
+  foldlMap1' = foldlMap1'PostOrder
+
+  foldlMap1 = foldlMap1PostOrder
+#endif
+
+foldrMap1PostOrder :: (a -> b) -> (a -> b -> b) -> PostOrder a -> b
+foldrMap1PostOrder f g = \(PostOrder (Node x ts)) ->
+  foldr (\t z -> foldr g z (PostOrder t)) (f x) ts
+{-# INLINE foldrMap1PostOrder #-}
+
+foldlMap1PostOrder :: (a -> b) -> (b -> a -> b) -> PostOrder a -> b
+foldlMap1PostOrder f g = \(PostOrder t) -> go t
+  where
+    go (Node x []) = f x
+    go (Node x (t:ts)) =
+      g (foldl (\z t' -> foldl g z (PostOrder t')) (go t) ts) x
+{-# INLINE foldlMap1PostOrder #-}
+
+foldlMap1'PostOrder :: (a -> b) -> (b -> a -> b) -> PostOrder a -> b
+foldlMap1'PostOrder f g = \(PostOrder t) -> go t
+  where
+    go (Node x []) = f x
+    go (Node x (t:ts)) =
+      let !z' = foldl' (\z t' -> foldl' g z (PostOrder t')) (go t) ts
+      in g z' x
+{-# INLINE foldlMap1'PostOrder #-}
+
 --------------------------------------------------------------------------------
 
 -- Note [Implemented Foldable Tree functions]
@@ -567,3 +849,38 @@ unfoldForestQ f aQ = case viewl aQ of
 --
 -- fold1, head: Defaults perform well.
 -- foldrMap1': Unlikely to be used.
+
+-- Note [pathsFromRoot implementation]
+-- ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+-- We use Okasaki's banker's queue for pathsFromRoot because it has some
+-- desirable properties when the result is consumed lazily.
+--
+-- 1. Fully evaluating a node's NonEmpty takes O(d) time, where d is
+--    the depth of the node. This is optimal.
+-- 2. The elements in the NonEmpty are yielded lazily. Note that the worst case
+--    time to yield an element is not O(1), i.e. it is only amortized O(1).
+--    More than O(1) work is done when the next element requires forcing (++)
+--    suspensions or reversing a rear list. For example, yielding the head has
+--    to force O(log d) (++) and so takes O(log d) time.
+-- 3. It builds up some beneficial sharing. It is not possible to share the
+--    results since the lists have different ends, but we can share some
+--    intermediate structures. Consider m sibling nodes at depth d. The front
+--    list is shared between them in (front ++ rear1), (front ++ rear2), ...
+--    (front + rearm). Forcing a prefix of front in one list can take arbitrary
+--    amounts of time per element (total bounded by O(d)), but once it is
+--    forced, front is memoized and doing the same for any of the siblings will
+--    take O(1) per element.
+--
+-- Alternatives:
+--
+-- * Implement it like pathsToRoot and reverse the NonEmptys. This does satisfy
+--   point 1 above. On 2 there's a trade-off, it costs a full O(d) to access the
+--   head and O(1) per element after that. On 3 it compares poorly because there
+--   is no sharing. Accessing the heads of m siblings will take O(dm) compared
+--   to the current O(d + m).
+-- * Use Okasaki's real-time queues. This would guarantee O(1) per element, but
+--   has worse constant-factor overall and does not seem worth the trouble.
+--
+-- GHC base also uses a banker's queue for Data.List.inits. inits is similar
+-- in nature to pathsFromRoot since a list is a tree where each node has one or
+-- zero children.